TR "Getting started"
Keep all the source in a single file. The way Rust handles modules and
dependencies is confusing me. Since this is a Tiny project, it should be
OK to keep everything in a single file. This allows me to move on,
without getting bogged down in making sense of it. If I ever use Rust in
a larger project, I am sure I will have the opportunity to learn how to
handle modules and crates and dependencies and what-not.

Export to Netpbm's Portable Pixel Map (PPM) format. The point of this
project is to make a rendered, not to implement an image format. PPM is
simple enough that I can define it myself, and not have to use an
external crate. By doing so, I get to explore a bit of Rust's IO.