TR Set pixels along Cartesian coordinates
`struct Image` should its origin at the top-left corner, to mirror how
screens produce images, and thus how the data is stored on disk.
However, the models we will be working with, use Cartesian coordinates.
To conciliate that, convert the coordinate system when setting a pixel.

Avoid using `panic!` in production, and ignore setting pixels that fall
outside the image.