TR "Bresenham’s line drawing algorithm"
Hardcode the path to the used model in `main.rs` instead of passing it
at runtime. The primary objective of this project is to create a
renderer, and not to parse a lot of OBJ files.

OBJ files in Sokolov's sample project are defined in the domain [-1, 1].
Those need to be transformed onto the [0,DIMENSION[ space before being
painted onto the image.

Avoid `panic!` in production by returning `Result` from `main()`.

`decode_obj()` does not adhere to the Wavefront .obj format, since it
specifies vertices can have 4 coordinates, and faces need not be
triangles. However, it is not necessary to accomodate those now.

`decode_obj()` does read through the passed string twice, which is not
optimal. However, it is fast enough, and it allows to define immutable
objects, so I am keeping it.