TR Implement `Vec2d`, `Vec3d`, and `Vec4d`
These are simply a type alias, so typing is not that strict, but it is
enough for now. It allows me to not have to re-implement underlying
methods, and call directly the indices.

But mostly, I wanted to try if this would work (it does), and it is
enough for this tiny render project.